updating logo image + renaming button on login page + adding more images
