+++ index.html
@@ -64,8 +64,8 @@
 
         <h1 class="purple-text" >v2.0.6</h1>
 
-        <a href="#" id="start_the_instalation" class="white-text btn waves-effect waves-light">sent msg to main.js</a>
-        <a href="#" id="download" class="white-text btn waves-effect waves-light purple hidden">download</a>
+        <a href="#" id="start_the_instalation" class="white-text btn waves-effect waves-light">INSTALL</a>
+        <a href="#" id="download" class="white-text btn waves-effect waves-light purple hidden">DOWNLOAD</a>
         <!-- bot nav -->
         <div class="row" style="position: absolute; bottom: 10px;">
           <!-- social icons -->
@@ -146,7 +146,7 @@
               <!-- action container -->
               <div class="row">
                 <!-- Login btn -->
-                <a id="btn_log_in" class="waves-effect waves-light btn" style="margin-left: 20px;">ΣΥΝΔΕΣΗ</a>
+                <a id="btn_log_in" class="waves-effect waves-light btn  light-blue darken-4" style="margin-left: 20px;">ΣΥΝΔΕΣΗ</a>
                 <br>
                 <br>
                 <!-- btn forget password -->
@@ -221,7 +221,7 @@
               <div class="row">
                 
                 <div class="col s6">
-                  <a id="btn_sign_in" class="waves-effect waves-light btn" style="color: white; margin-left: 10px; margin-top: 10px;">SING IN</a>
+                  <a id="btn_sign_in" class="waves-effect waves-light btn light-blue darken-4" style="color: white; margin-left: 10px; margin-top: 10px;">SING IN</a>
                 </div>
                 <div class="col s6">
                   <a id="undo_singUp" class="waves-effect waves-light btn white" style="color: black; margin-right: 0px; margin-top: 10px;">BACK</a>
@@ -255,8 +255,9 @@
 
 
   <script src="./Pages/updater.js" type="text/javascript"></script>
+  <script src="./Pages/authentication.js" type="text/javascript"></script>
   <script src="./Pages/login.js" type="text/javascript"></script>
-  <!--<script src="index.js" type="text/javascript"></script>-->
+  <script src="./Pages/index.js" type="text/javascript"></script>
 </body>
 
 </html>
