+++ main.css
@@ -56,8 +56,8 @@ body {
   width: calc(100vw - 300px);
   height: 100vh;
   /*background: green;*/
-    background-image: url('Assets/img/mykonos.jpg');
-    background-position:bottom;
+    background-image: url('Assets/img/WelcomePage/greenPurplw.jpg');
+    background-position:top;
 
 }
